pyproject.toml
[project]
name = "sdkportfolio-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.6",
    "openai-agents>=0.2.1",
    "pinecone>=7.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
]
